ResourceGroupsTaggingAPI tests - move into dedicated file
